+++ cucumber/ios/features/step_definitions/shared_steps.rb
@@ -1,14 +1,14 @@
-Given(/^I see the (first|second|third|gestures) tab$/) do |tab|
+Given(/^I see the (first|scrolls|special|tapping) tab$/) do |tab|
   wait_for_view('tabBarButton')
   case tab
-    when 'first'
-      index = 0
-    when 'second'
-      index = 1
-    when 'third'
-      index = 2
-    when 'gestures'
-      index = 3
+  when 'first'
+    index = 0
+  when 'scrolls'
+    index = 1
+  when 'special'
+    index = 2
+  when 'tapping'
+    index = 3
   end
   tap("tabBarButton index:#{index}")
   expected_view = "#{tab} page"
